Enhance networking modules with IPv6 support and integration
- Added frontdesk and concierge modules for IPv6 address management.
- Updated Makefile to include Networking directory in source search.
- Refactored FrontDoor and FrontDesk for improved networking functionality.
- Implemented periodic maintenance in timer callback for networking.
- Introduced waystation_drain function for better connection handling.

+++ kernel/PCI/pci.h
@@ -1,4 +1,5 @@
 #ifndef PCI_H
+
 #define PCI_H
 
 #include <stdint.h>
@@ -14,7 +15,7 @@ typedef struct {
     int found;
 }pci_device_t;
 
-uint32_t pci_config_read32(uint32_t bus,uint8_t slot,uint8_t func,uint8_t offset);
+uint32_t pci_config_read32(uint8_t bus,uint8_t slot,uint8_t func,uint8_t offset);
 uint16_t pci_config_read16(uint8_t bus,uint8_t slot,uint8_t func,uint8_t offset);
 uint8_t pci_config_read8(uint8_t bus,uint8_t slot,uint8_t func,uint8_t offset);
 void pci_config_write32(uint8_t bus, uint8_t slot, uint8_t func, uint8_t offset, uint32_t value);
